main/requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
async-timeout==5.0.1
asyncio-mqtt==0.16.2
asyncpg==0.30.0
av==14.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
email-validator==2.3.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
ffmpeg-python==0.2.0
future==1.0.0
cffi==2.0.0
circuitbreaker==2.1.3
click==8.1.8
cryptography==46.0.1
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.118.0
fastapi-cache2==0.2.2
Flask==3.1.2
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
jmespath==0.10.0
MyApplication==0.1.0
paho-mqtt==2.1.0
passlib==1.7.4
pendulum==3.1.0
prompt-toolkit==3.0.43
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.3.0
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
PyYAML==6.0.3
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
terminaltables==3.1.10
tortoise-orm==0.25.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0
